Zips of some P5.js art
I was attempting to make a square appear every few seconds that is a random color. It took a few iterations for me to do it the way i wanted. The file called 'SuaveSea' is my first attempt - the sqaures were too fast- so I added the setInterval but I had it wrong. After googling it came out like Colourful Squares